:bug: Fix at::server_write() memory bug
server_write was supposed to clamp the input to the maximum send limit,
but instead, extended the span to the maximum send limit, spilling out
contents of rodata to the server which is bad on so many levels.

Updated demo to use a streamed based class for parsing an HTTP get
request.